Docs, fixed Schema, adding Bus real-time config

+++ src/entities/position.entity.ts
@@ -1,11 +1,11 @@
-import { Field, Int, ObjectType } from '@nestjs/graphql';
+import { Field, Float, Int, ObjectType } from '@nestjs/graphql';
 
 @ObjectType()
 export class PositionEntity {
-  @Field(() => Int, { nullable: true })
+  @Field(() => Float, { nullable: true })
   latitude?: number;
 
-  @Field(() => Int, { nullable: true })
+  @Field(() => Float, { nullable: true })
   longitude?: number;
 
   @Field(() => Int, { nullable: true })
